# lib/api-zod/src/generated/types/createNoteBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EduPath Student Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateNoteBody {
  title: string;
  subject: string;
  folder: string;
  educationLevel?: string;
  description?: string;
  fileUrl?: string;
  fileType?: string;
  tags?: string[];
}