# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EduPath Student Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { DashboardSummaryUpcomingDeadlinesItem } from "./dashboardSummaryUpcomingDeadlinesItem";

export interface DashboardSummary {
  totalExams: number;
  openExams: number;
  totalJobs: number;
  openJobs: number;
  totalNotes: number;
  totalCareers: number;
  upcomingDeadlines: DashboardSummaryUpcomingDeadlinesItem[];
  latestNotifications: number;
  aiTrends: number;
}